구현 : 지도 값 불러오기 스캘래톤 코드 - json.dart로 API 연결 함수 몰아 넣기 Issue #107
